# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scotland Yard Multiplayer Game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type CreateGameBodyRole =
  (typeof CreateGameBodyRole)[keyof typeof CreateGameBodyRole];

export const CreateGameBodyRole = {
  mrx: "mrx",
  detective: "detective",
} as const;

export interface CreateGameBody {
  playerName: string;
  role: CreateGameBodyRole;
}

export type JoinGameBodyRole =
  (typeof JoinGameBodyRole)[keyof typeof JoinGameBodyRole];

export const JoinGameBodyRole = {
  mrx: "mrx",
  detective: "detective",
} as const;

export interface JoinGameBody {
  playerName: string;
  role: JoinGameBodyRole;
}

export interface StartGameBody {
  playerId: string;
}

export type MakeMoveBodyTransport =
  (typeof MakeMoveBodyTransport)[keyof typeof MakeMoveBodyTransport];

export const MakeMoveBodyTransport = {
  taxi: "taxi",
  bus: "bus",
  underground: "underground",
  black: "black",
  boat: "boat",
} as const;

export interface MakeMoveBody {
  playerId: string;
  toStation: number;
  transport: MakeMoveBodyTransport;
  useDouble?: boolean;
}

export type PlayerRole = (typeof PlayerRole)[keyof typeof PlayerRole];

export const PlayerRole = {
  mrx: "mrx",
  detective: "detective",
} as const;

export interface Tickets {
  taxi: number;
  bus: number;
  underground: number;
  black: number;
  double: number;
}

export interface Player {
  id: string;
  name: string;
  role: PlayerRole;
  /** @nullable */
  position: number | null;
  tickets: Tickets;
  isConnected: boolean;
  doubleMoveActive?: boolean;
  emoji?: string;
}

export interface MrXMove {
  round: number;
  transport: string;
  /** @nullable */
  station: number | null;
  isDoubleMove?: boolean;
}

export type GameStatus = (typeof GameStatus)[keyof typeof GameStatus];

export const GameStatus = {
  lobby: "lobby",
  playing: "playing",
  mrx_won: "mrx_won",
  detectives_won: "detectives_won",
} as const;

export interface Game {
  id: string;
  roomCode: string;
  status: GameStatus;
  players: Player[];
  /** @nullable */
  currentTurn: string | null;
  round: number;
  maxRounds: number;
  mrxRevealRounds: number[];
  mrxMoveLog: MrXMove[];
  turnOrder: string[];
  turnIndex: number;
  /** @nullable */
  winner: string | null;
  /** @nullable */
  mrxStartQuarter?: number | null;
  /** @nullable */
  decoyQuarter?: number | null;
}