# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scotland Yard Multiplayer Game API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create a new game room
 */
export const CreateGameBody = zod.object({
  playerName: zod.string(),
  role: zod.enum(["mrx", "detective"]),
});

/**
 * @summary Get game state
 */
export const GetGameParams = zod.object({
  roomCode: zod.coerce.string(),
});

const playerSchema = zod.object({
  id: zod.string(),
  name: zod.string(),
  role: zod.enum(["mrx", "detective"]),
  position: zod.number().nullable(),
  tickets: zod.object({
    taxi: zod.number(),
    bus: zod.number(),
    underground: zod.number(),
    black: zod.number(),
    double: zod.number(),
  }),
  isConnected: zod.boolean(),
  doubleMoveActive: zod.boolean().optional(),
});

const mrxMoveSchema = zod.object({
  round: zod.number(),
  transport: zod.string(),
  station: zod.number().nullable(),
  isDoubleMove: zod.boolean().optional(),
});

const gameResponseSchema = zod.object({
  id: zod.string(),
  roomCode: zod.string(),
  status: zod.enum(["lobby", "playing", "mrx_won", "detectives_won"]),
  players: zod.array(playerSchema),
  currentTurn: zod.string().nullable(),
  round: zod.number(),
  maxRounds: zod.number(),
  mrxRevealRounds: zod.array(zod.number()),
  mrxMoveLog: zod.array(mrxMoveSchema),
  turnOrder: zod.array(zod.string()),
  turnIndex: zod.number(),
  winner: zod.string().nullable(),
});

export const GetGameResponse = gameResponseSchema;

/**
 * @summary Join a game room
 */
export const JoinGameParams = zod.object({
  roomCode: zod.coerce.string(),
});

export const JoinGameBody = zod.object({
  playerName: zod.string(),
  role: zod.enum(["mrx", "detective"]),
});

export const JoinGameResponse = gameResponseSchema;

/**
 * @summary Start the game
 */
export const StartGameParams = zod.object({
  roomCode: zod.coerce.string(),
});

export const StartGameBody = zod.object({
  playerId: zod.string(),
});

export const StartGameResponse = gameResponseSchema;

/**
 * @summary Make a move
 */
export const MakeMoveParams = zod.object({
  roomCode: zod.coerce.string(),
});

export const MakeMoveBody = zod.object({
  playerId: zod.string(),
  toStation: zod.number(),
  transport: zod.enum(["taxi", "bus", "underground", "black", "boat"]),
  useDouble: zod.boolean().optional(),
});

export const MakeMoveResponse = gameResponseSchema;