# lib/api-zod/src/generated/types/createGameBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scotland Yard Multiplayer Game API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateGameBodyRole } from "./createGameBodyRole";

export interface CreateGameBody {
  playerName: string;
  role: CreateGameBodyRole;
}