# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scotland Yard Multiplayer Game API
 * OpenAPI spec version: 0.1.0
 */
import type { GameStatus } from "./gameStatus";
import type { MrXMove } from "./mrXMove";
import type { Player } from "./player";

export interface Game {
  id: string;
  roomCode: string;
  status: GameStatus;
  players: Player[];
  /** @nullable */
  currentTurn: string | null;
  round: number;
  maxRounds: number;
  mrxRevealRounds: number[];
  mrxMoveLog: MrXMove[];
  turnOrder: string[];
  turnIndex: number;
  /** @nullable */
  winner: string | null;
  /** @nullable */
  mrxStartQuarter?: number | null;
  /** @nullable */
  decoyQuarter?: number | null;
}