# lib/api-zod/src/generated/types/joinGameBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scotland Yard Multiplayer Game API
 * OpenAPI spec version: 0.1.0
 */
import type { JoinGameBodyRole } from "./joinGameBodyRole";

export interface JoinGameBody {
  playerName: string;
  role: JoinGameBodyRole;
}