# lib/api-zod/src/generated/types/makeMoveBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scotland Yard Multiplayer Game API
 * OpenAPI spec version: 0.1.0
 */
import type { MakeMoveBodyTransport } from "./makeMoveBodyTransport";

export interface MakeMoveBody {
  playerId: string;
  toStation: number;
  transport: MakeMoveBodyTransport;
  useDouble?: boolean;
}