# lib/api-zod/src/generated/types/mrXMove.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scotland Yard Multiplayer Game API
 * OpenAPI spec version: 0.1.0
 */

export interface MrXMove {
  round: number;
  transport: string;
  /** @nullable */
  station: number | null;
  isDoubleMove?: boolean;
}