# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scotland Yard Multiplayer Game API
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerRole } from "./playerRole";
import type { Tickets } from "./tickets";

export interface Player {
  id: string;
  name: string;
  role: PlayerRole;
  /** @nullable */
  position: number | null;
  tickets: Tickets;
  isConnected: boolean;
  doubleMoveActive?: boolean;
}